everything ported but file access

+++ Vertex.cs
@@ -9,6 +9,6 @@ namespace csg_NET
     public class Vertex
     {
         public Vector3 p = Vector3.zero;
-        public double[] tex = new double[2];
+        public float[] tex = new float[2];
     }
 }
